mirai-agent/requirements.txt
# Mirai Autonomous Agent - Dependencies

# Core AI / HTTP clients
openai>=1.0.0
requests>=2.31.0
httpx>=0.28.0
uvicorn>=0.30
fastapi>=0.115

# Data/config handling
PyYAML>=6.0
aiohttp>=3.13.0
numpy>=1.26
pandas>=2.2
psutil>=5.9
prometheus-client>=0.20

# Async caching / messaging
aioredis>=2.0
websockets>=11.0

# Notifications & scheduling
python-telegram-bot>=20.0
schedule>=1.2
email-validator>=2.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.9
bcrypt==4.1.2

# Windows Service (optional)
pywin32>=300; platform_system == 'Windows'
